+++ BWAPI4J/src/main/java/org/openbw/bwapi4j/Position.java
@@ -20,7 +20,7 @@
 
 package org.openbw.bwapi4j;
 
-import static org.openbw.bwapi4j.util.MathUtil.estimateDistanceBetween;
+import static org.openbw.bwapi4j.util.MathUtils.estimateDistanceBetween;
 
 public class Position {
